% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleProfile2metdna.R
\name{sampleProfile2metdna}
\alias{sampleProfile2metdna}
\title{sampleProfile2metdna}
\usage{
sampleProfile2metdna(path = ".", polarity = c("positive", "negative"),
  width = 10, height = 6)
}
\arguments{
\item{path}{Absolute directory}

\item{polarity}{polarity}

\item{width}{width, in}

\item{height}{height, in}
}
\value{
A png figure
}
\description{
sampleProfile2metdna
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
